requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
pydantic-settings
dependency_injector
openai
python-toon
redis
psycopg2-binary